Add a setting for a connection pause
Connection pause waits a specified number of seconds after opening the
port to the printer before commencing serial communication.  The default
in GPX before the plugin project was 2 seconds, so that remains after
this commit. I've tested my flashforge creator and it doesn't seem to
need much time (0.5 seconds works reliably for example).