Remove requirement on cancel for a reconnect
Up to this point we relied on a disconnect/connect after throwing an
error on cancel.  Now we have to distinguish between cancel's initiated
by OctoPrint and those initiated on the LCD panel.

Why do we want to avoid disconnect/connect?  Well, since the bot always
returns the cancel build status first on an overtemp condition we have
to stay connected to get to hear the overtemp in case we want to take
some action in OctoPrint (like turn off a relay for example)